Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -12,7 +12,7 @@ int main() {
     else if (c == '*')
         printf("%d", a * b);
     else if (c == '/')
-        printf("%f", a / b);
+        printf("%f", (float)a / b);
     else
         printf("Error");
 }
